+++ angular/src/lib/input/examples/examples.module.ts
@@ -1,6 +1,9 @@
 import { NgModule, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
+
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { InputModule } from '../input.module';
+
+import { InputModule } from '@momentum-ui/angular';
+
 import { CommonModule } from '@angular/common';
 
 import {