+++ vue/src/examples.js
@@ -1,3 +1,5 @@
+import Alert from './lib/alert/examples/index.js';
+import AlertBanner from './lib/alert-banner/examples/index.js';
 import Badge from './lib/badge/examples/index.js';
 import Button from './lib/button/examples/index.js';
 import ButtonGroup from './lib/button-group/examples/index.js';
@@ -9,9 +11,13 @@ import List from './lib/list/examples/index.js';
 import ListItem from './lib/list-item/examples/index.js';
 import ListItemHeader from './lib/list-item-header/examples/index.js';
 import Loading from './lib/loading/examples/index.js';
+import Modal from './lib/modal/examples/index.js';
+import Popover from './lib/popover/examples/index.js';
 import Radio from './lib/radio/examples/index.js';
 
 const examples = [
+  Alert,
+  AlertBanner,
   Badge,
   Button,
   ButtonGroup,
@@ -23,6 +29,8 @@ const examples = [
   ListItem,
   ListItemHeader,
   Loading,
+  Modal,
+  Popover,
   Radio
 ];