+++ vue/src/lib/modal-body/index.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="md-modal__body">
+    <slot></slot>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'md-modal-body',
+};
+</script>