+++ angular/src/lib/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.3.1](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/angular@8.3.0...@momentum-ui/angular@8.3.1) (2019-09-11)
+
+**Note:** Version bump only for package @momentum-ui/angular
+
+
+
+
+
+# [8.3.0](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/angular@8.2.8...@momentum-ui/angular@8.3.0) (2019-09-07)
+
+
+### Features
+
+* **radio:** Add support for helpText ([8463a4b](https://github.com/momentum-design/momentum-ui/commit/8463a4b))
+
+
+
+
+
+## [8.2.8](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/angular@8.2.7...@momentum-ui/angular@8.2.8) (2019-09-06)
+
+**Note:** Version bump only for package @momentum-ui/angular
+
+
+
+
+
+## [8.2.7](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/angular@8.2.6...@momentum-ui/angular@8.2.7) (2019-09-05)
+
+**Note:** Version bump only for package @momentum-ui/angular
+
+
+
+
+
+## [8.2.6](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/angular@8.2.5...@momentum-ui/angular@8.2.6) (2019-08-30)
+
+
+### Bug Fixes
+
+* **input:** revert breaking change ([7249803](https://github.com/momentum-design/momentum-ui/commit/7249803))
+
+
+
+
+
 ## [8.2.5](https://github.com/momentum-design/momentum-ui/compare/@momentum-ui/angular@8.2.4...@momentum-ui/angular@8.2.5) (2019-08-29)